+++ examples/shell_scripts/prune_failing_machines_signals.py
@@ -4,9 +4,17 @@ This script deletes signals linked to a failing machine.
 
 import argparse
 import sys
+import logging
 
 from cscapi.client import CAPIClient, CAPIClientConfig
 from cscapi.sql_storage import SQLStorage
+
+logger = logging.getLogger("capi-py-sdk")
+logger.setLevel(logging.DEBUG)  # Change this to the level you want
+console_handler = logging.StreamHandler()
+formatter = logging.Formatter("%(asctime)s - %(name)s - %(levelname)s - %(message)s")
+console_handler.setFormatter(formatter)
+logger.addHandler(console_handler)
 
 
 class CustomHelpFormatter(argparse.HelpFormatter):
@@ -26,6 +34,12 @@ try:
         help="Local database name. Example: cscapi.db",
         required=True,
     )
+    parser.add_argument(
+        "--batch_size",
+        type=int,
+        help="Batch size for pruning signals. Example: 1000",
+        default=1000,
+    )
     args = parser.parse_args()
 except argparse.ArgumentError as e:
     print(e)
@@ -34,11 +48,13 @@ except argparse.ArgumentError as e:
 
 database = args.database
 database_message = f"\tLocal storage database: {database}\n"
+batch_size_message = f"\tBatch size: {args.batch_size}\n"
 
 print(
     f"\nPruning signals for failing machines\n\n"
     f"Details:\n"
     f"{database_message}"
+    f"{batch_size_message}"
     f"\n\n"
 )